ai_scout_batch_2026_04_07/ai-agent-persistent-context-memory/requirements.txt
SQLAlchemy
chromadb
sentence-transformers
openai
tiktoken
networkx
pydantic
python-dotenv